encoded_id列の変換がfloat型になっていたのでint型に修正

+++ modules/preprocessing/_feature_engineering.py
@@ -83,7 +83,7 @@ class FeatureEngineering:
             target_master = pd.read_csv(csv_path, dtype=object)
 
         # 後のmaxでエラーになるので、整数に変換
-        target_master['encoded_id'] = target_master['encoded_id'].astype(float)
+        target_master['encoded_id'] = target_master['encoded_id'].astype(int)
 
         # masterに存在しない、新しい情報を抽出
         new_target = self.__data[[target_col]][
